+++ src/game/SocialMgr.h
@@ -1,4 +1,22 @@
-
+/*
+ * 
+ *
+ * 
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+ */
 
 #ifndef __QUAD_SOCIALMGR_H
 #define __QUAD_SOCIALMGR_H
@@ -33,8 +51,8 @@ struct FriendInfo
     FriendStatus Status;
     uint32 Flags;
     uint32 Area;
-    uint32 Level;
-    uint32 Class;
+    uint8 Level;
+    uint8 Class;
     std::string Note;
 
     FriendInfo()
